refactor(api): rename ambiguous public modules (atomic->atomic_spans, ast->markdown_ast)
"atomic" and "ast" are overloaded single words (and "ast" shadows the stdlib
module). Rename for unambiguous, self-describing public import paths:
flowmark.atomic -> flowmark.atomic_spans, flowmark.ast -> flowmark.markdown_ast.
Pure rename; all references (code, tests, spec) updated.
